requirements.txt
flask
gunicorn
waitress
python-dotenv
scikit-learn
openai
langchain
flask-cors
pydantic==1.10.9
chainfury>=1.4.3